chrome/content/mochitest/tests/test_requestList_memory.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tests for request data</title>

  <link rel="stylesheet" type="text/css" href="/content/tests/SimpleTest/test.css" />

  <script type="text/javascript" src="/content/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/x-javascript;version=1.7" src="../httpd.js"></script>
  <script type="application/x-javascript;version=1.7" src="common.js"></script>
</head>
<body>
  <p id="display"></p>
  <div id="content" style="display: none">
    <iframe id="frame"></iframe>
  </div>
  <pre id="test">
  <script type="application/x-javascript;version=1.7">
    let abp = Components.classes["@mozilla.org/adblockplus;1"].createInstance().wrappedJSObject;

    function start()
    {
      window.addEventListener("unload", stop, false);
      server.start(1234);

      abp.prefs.enabled = true;

      server.registerPathHandler("/test1", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<script>for (var i = 0; i < 40; i++) new Image().src='test.gif';<" + "/script>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test2", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<body onload='document.body.innerHTML=null;'><script>for (var i = 0; i < 40; i++) document.write('<iframe src=dummy></iframe>')<" + "/script></body>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test3", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<body><script>for (var i = 0; i < 40; i++) document.write('<iframe src=dummy></iframe>')<" + "/script></body>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test4", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<script>function addImage() { new Image().src='test.gif'; }for (var i = 0; i < 40; i++) addImage();<" + "/script>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test5", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<script>for (var i = 0; i < 40; i++) new Image().src='test'+i+'.gif';<" + "/script>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test6", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<body onload='document.body.innerHTML=null;'><script>for (var i = 0; i < 40; i++) document.write('<iframe src=dummy'+i+'></iframe>')<" + "/script></body>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test7", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<body><script>for (var i = 0; i < 40; i++) document.write('<iframe src=dummy' +i+'></iframe>')<" + "/script></body>";
        response.bodyOutputStream.write(body, body.length);
      });

      server.registerPathHandler("/test8", function(metadata, response)
      {
        response.setStatusLine("1.1", "200", "OK");
        response.setHeader("Content-Type", "text/html");

        let body = "<script>function addImage() { new Image().src='test'+i+'.gif'; }for (var i = 0; i < 40; i++) addImage();<" + "/script>";
        response.bodyOutputStream.write(body, body.length);
      });

      $("frame").onload = function() setTimeout(test1Done, 0);;
      $("frame").src = "http://localhost:1234/test1";
    }

    function test1Done()
    {
      let frameData = abp.RequestList.getDataForWindow(abp.wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data");

      let entry = entries[0];
      is(entry.nodes.length, 40, "Number of associated nodes before garbage collection");

      window.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      is(entry.nodes.length, 0, "Number of associated nodes after garbage collection");

      $("frame").onload = function() setTimeout(test2Done, 0);;
      $("frame").src = "http://localhost:1234/test2";
    }

    function test2Done()
    {
      let frameData = abp.RequestList.getDataForWindow(abp.wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data");

      let entry = entries[0];
      is(entry.nodes.length, 40, "Number of associated nodes before garbage collection");

      window.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      is(entry.nodes.length, 0, "Number of associated nodes after garbage collection");

      $("frame").onload = function() setTimeout(test3Done, 0);;
      $("frame").src = "http://localhost:1234/test3";
    }

    function test3Done()
    {
      let frameData = abp.RequestList.getDataForWindow(abp.wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data");

      let entry = entries[0];
      is(entry.nodes.length, 40, "Number of associated nodes before garbage collection");

      window.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      is(entry.nodes.length, 40, "Number of associated nodes after garbage collection");

      $("frame").onload = function() setTimeout(test4Done, 0);
      $("frame").src = "http://localhost:1234/test4";
    }

    function test4Done(event)
    {
      let frameData = abp.RequestList.getDataForWindow(abp.wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data");

      let entry = entries[0];
      is(entry._nodes.length, 40, "Number of associated nodes before garbage collection");

      window.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();

      // Trigger list compression by adding a node while the counter has a high value
      entry._compactCounter = 100;
      window.frames[0].addImage();
      is(entry._nodes.length, 1, "Number of associated nodes after garbage collection");

      $("frame").onload = function() setTimeout(test5Done, 0);;
      $("frame").src = "http://localhost:1234/test5";
    }

    function test5Done()
    {
      let frameData = abp.RequestList.getDataForWindow(abp.wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data before garbage collection");

      window.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      for each (let entry in entries)
        entry.lastUpdate = 0;

      entries = frameData.getAllLocations();
      is(entries.length, 0, "Number of entries in the frame data after garbage collection");

      $("frame").onload = function() setTimeout(test6Done, 0);;
      $("frame").src = "http://localhost:1234/test6";
    }

    function test6Done()
    {
      let frameData = abp.RequestList.getDataForWindow(abp.wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data before garbage collection");

      window.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      for each (let entry in entries)
        entry.lastUpdate = 0;

      entries = frameData.getAllLocations();
      is(entries.length, 0, "Number of entries in the frame data after garbage collection");

      $("frame").onload = function() setTimeout(test7Done, 0);;
      $("frame").src = "http://localhost:1234/test7";
    }

    function test7Done()
    {
      let frameData = abp.RequestList.getDataForWindow(abp.wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data before garbage collection");

      window.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      for each (let entry in entries)
        entry.lastUpdate = 0;

      entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data after garbage collection");

      $("frame").onload = function() setTimeout(test8Done, 0);;
      $("frame").src = "http://localhost:1234/test8";
    }

    function test8Done()
    {
      let frameData = abp.RequestList.getDataForWindow(abp.wrapNode(window.frames[0]), true);
      ok(frameData, "Frame has data associated with it");

      let entries = frameData.getAllLocations();
      is(entries.length, 40, "Number of entries in the frame data before garbage collection");

      window.QueryInterface(Ci.nsIInterfaceRequestor).getInterface(Ci.nsIDOMWindowUtils).garbageCollect();
      for each (let entry in entries)
        entry.lastUpdate = 0;

      // Trigger list compression by adding a node while the counter has a high value
      frameData._compactCounter = 100;
      window.frames[0].addImage();

      entries = frameData.getAllLocations();
      is(entries.length, 1, "Number of entries in the frame data after garbage collection");

      $("frame").onload = null;
      SimpleTest.finish();
    }

    function stop()
    {
      server.stop();

      abp.prefs.enabled = oldEnabled;
    }

    let server = new nsHttpServer();
    let oldEnabled = abp.prefs.enabled;
    SimpleTest.waitForExplicitFinish();
    addLoadEvent(start);
  </script>
  </pre>
</body>
</html>
